Trekk ut felles ekspanderbart panel
